Updates include: - add SearchCard/SearchCardBody/ folder - create the sub-component  to render each section on the SearchCardBody component - (minor) add the edge case(for empty string) to helpers/getURLForAccessionCode
